requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
httpx>=0.25.0
pydantic>=2.0.0
python-multipart>=0.0.6
joblib>=1.3.0
argon2-cffi>=23.1.0
pyjwt>=2.8.0